requirements.txt
pylocres
opencc-python-reimplemented
